pc - allow CSV upload of GitHub ids
In this commit, we add code that allows CSV files to have a GitHub id column.
If filled in, and if the GitHub id is valid, it will add the github id to the Roster Student
record.  This allows CSV download/upload to use used to migrate a course from one
instance of the app to another.

If the user has never logged in, the user record is created in the user table.