pc - expand graphql query for issue sdlc events
This is everything that looks reasonable to get on a first pass
for events.  It includes creating, closing, reopening,
adding to project, moving within project, removing from project,
as well as edits to the issue itself.

It does NOT currently include commenting on the issue,
or adding attachments.  There are still more things we could get
but this is a good start.