pc - refactored reporting code for jobs and got more SDLC events
Put infrastructure in place to get issue events and PR events separately
but in the same job for simplicity.   Database support not yet done;
need to think through the design since there is a lot of variation in
the data generated for events.  Probably should separate issues and PRs,
but maybe editing events are something they share in common?

Should it be a MongoDB, but ouch that's a lot to take on; getting all that
new infrastructure in place.    Maybe better to make it work with SQL for
now.