pc - refactor tech debt in testing
We reduce some of the flakiness by ensuring that tests don't reassign @user instance
variable in the course_controller_tests.  We also restore the relationships that were
broken in the fixtures (and that the old minitest rails tests rely on), and
that we broke by assiging fixed ids to courses.   We then find a different way,
a more robust way, to handle getting a specific course url in the cypress tests;
we save the first course that we find on the rendered page, and use that in the
subsequent tests.