pc - add test for mismatch in email case
A bug was reported by an instructor using the app, where a student's emai
email address on the course roster was in uppercase, but in practice
they always used lowercase.  In truth, email addresses are typically
case insensitive.   This test covers that.

It is written in the style of the tests currently in the test suite,
which is NOT the best style; it relies on fixtures far from the tests
themselves.  Refactoring that is a larger project; for now, we just needed
to get a test in to cover this case.