pc/at - encode the name and or email from the GitHub commit data
Otherwise, rails may try to 'unescape' or 'decode' this data when it is passed as a URL parameter,
and that may result in incorrect/unexpected results.  For example, sometimes the name and or email has a +
character in it, which rails, incorrectly, treats as an escaped space character and will try to change
back into a space.