pc - can define multiple exclusion criteria
In the method  file_should_be_excluded in app/jobs/repo/course_github_repo_get_commits.rb

Note that this logic should eventually be factored out so that it applies to incoming
GitHub event triggers as well.    For now, it only works when we run the job to get the
commits.